+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - "user" variable name for a character not replaced by the player name playing the story
 - Default ground of entity to false
 - Gameplay tag not automatically show hud again
+- Quote inside quote not correctly validated on tags